move most config to separate class
